Move excluding by target labels to generic file finding module
