+++ mbed_build/_internal/mbed_lib.py
@@ -6,12 +6,10 @@
 from pathlib import Path
 from typing import Iterable
 
-from mbed_targets import get_build_attributes_by_board_type
-
 from mbed_build._internal.find_files import (
     find_files,
     exclude_using_mbedignore,
-    exclude_using_labels,
+    exclude_using_target_labels,
 )
 
 
@@ -24,14 +22,5 @@ def find_mbed_lib_files(mbed_program_directory: Path, board_type: str) -> Iterab
     """
     mbed_lib_paths = find_files("mbed_lib.json", mbed_program_directory)
     mbed_lib_paths = exclude_using_mbedignore(mbed_program_directory, mbed_lib_paths)
-    mbed_lib_paths = _exclude_using_target_labels(mbed_program_directory, board_type, mbed_lib_paths)
+    mbed_lib_paths = exclude_using_target_labels(mbed_program_directory, board_type, mbed_lib_paths)
     return mbed_lib_paths
-
-
-def _exclude_using_target_labels(mbed_program_directory, board_type, paths):
-    """Filter out given paths using target labels."""
-    build_attributes = get_build_attributes_by_board_type(board_type, mbed_program_directory)
-    paths = exclude_using_labels("TARGET", build_attributes.labels, paths)
-    paths = exclude_using_labels("FEATURE", build_attributes.features, paths)
-    paths = exclude_using_labels("COMPONENT", build_attributes.components, paths)
-    return paths
